# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IORMCreateResponse": {
				"description": "Réponse à une operation d'insertion d'une ligne",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "ID de la ligne créée"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUser": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IORMIndexResponse_IUser_": {
				"description": "Réspone à une opération de lecture de plusieurs lignes.",
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IUser"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.Exclude_keyofIUser.userId__": {
				"properties": {
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUser.userId_": {
				"$ref": "#/components/schemas/Pick_IUser.Exclude_keyofIUser.userId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IUserCreate": {
				"$ref": "#/components/schemas/Omit_IUser.userId_"
			},
			"IORMUpdateResponse": {
				"description": "Réponse à une operation de mise à jour",
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"rows": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IUserCreate_": {
				"properties": {
					"familyName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IUserUpdate": {
				"$ref": "#/components/schemas/Partial_IUserCreate_"
			},
			"IORMDeleteResponse": {
				"description": "Réponse à une operation de suppression d'une ligne",
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"rows": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IFilm": {
				"properties": {
					"film_id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"release_year": {
						"type": "number",
						"format": "double"
					},
					"language_id": {
						"type": "number",
						"format": "double"
					},
					"original_language_id": {
						"type": "number",
						"format": "double"
					},
					"rental_duration": {
						"type": "number",
						"format": "double"
					},
					"rental_rate": {
						"type": "number",
						"format": "double"
					},
					"length": {
						"type": "number",
						"format": "double"
					},
					"replacement_cost": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "string",
						"enum": [
							"G",
							"PG",
							"PG-13",
							"R",
							"NC-17"
						]
					},
					"special_features": {
						"items": {
							"type": "string",
							"enum": [
								"Trailers",
								"Commentaries",
								"Deleted Scenes",
								"Behind the Scenes"
							]
						},
						"type": "array"
					},
					"last_update": {
						"type": "string"
					}
				},
				"required": [
					"film_id",
					"title",
					"language_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IORMIndexResponse_IFilm_": {
				"description": "Réspone à une opération de lecture de plusieurs lignes.",
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IFilm"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IFilm.Exclude_keyofIFilm.film_id__": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"release_year": {
						"type": "number",
						"format": "double"
					},
					"language_id": {
						"type": "number",
						"format": "double"
					},
					"original_language_id": {
						"type": "number",
						"format": "double"
					},
					"rental_duration": {
						"type": "number",
						"format": "double"
					},
					"rental_rate": {
						"type": "number",
						"format": "double"
					},
					"length": {
						"type": "number",
						"format": "double"
					},
					"replacement_cost": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "string",
						"enum": [
							"G",
							"PG",
							"PG-13",
							"R",
							"NC-17"
						]
					},
					"special_features": {
						"items": {
							"type": "string",
							"enum": [
								"Trailers",
								"Commentaries",
								"Deleted Scenes",
								"Behind the Scenes"
							]
						},
						"type": "array"
					},
					"last_update": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"language_id"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IFilm.film_id_": {
				"$ref": "#/components/schemas/Pick_IFilm.Exclude_keyofIFilm.film_id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IFilmCreate": {
				"$ref": "#/components/schemas/Omit_IFilm.film_id_"
			},
			"Partial_IFilmCreate_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"release_year": {
						"type": "number",
						"format": "double"
					},
					"language_id": {
						"type": "number",
						"format": "double"
					},
					"original_language_id": {
						"type": "number",
						"format": "double"
					},
					"rental_duration": {
						"type": "number",
						"format": "double"
					},
					"rental_rate": {
						"type": "number",
						"format": "double"
					},
					"length": {
						"type": "number",
						"format": "double"
					},
					"replacement_cost": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "string",
						"enum": [
							"G",
							"PG",
							"PG-13",
							"R",
							"NC-17"
						]
					},
					"special_features": {
						"items": {
							"type": "string",
							"enum": [
								"Trailers",
								"Commentaries",
								"Deleted Scenes",
								"Behind the Scenes"
							]
						},
						"type": "array"
					},
					"last_update": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IFilmUpdate": {
				"$ref": "#/components/schemas/Partial_IFilmCreate_"
			},
			"IActor": {
				"properties": {
					"actor_id": {
						"type": "number",
						"format": "double"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"last_update": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"actor_id",
					"first_name",
					"last_name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IORMIndexResponse_IActor_": {
				"description": "Réspone à une opération de lecture de plusieurs lignes.",
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IActor"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IActor.Exclude_keyofIActor.actor_id__": {
				"properties": {
					"last_update": {
						"type": "string",
						"format": "date-time"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					}
				},
				"required": [
					"first_name",
					"last_name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IActor.actor_id_": {
				"$ref": "#/components/schemas/Pick_IActor.Exclude_keyofIActor.actor_id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IActorCreate": {
				"$ref": "#/components/schemas/Omit_IActor.actor_id_"
			},
			"Partial_IActorCreate_": {
				"properties": {
					"last_update": {
						"type": "string",
						"format": "date-time"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IActorUpdate": {
				"$ref": "#/components/schemas/Partial_IActorCreate_"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "jwt",
				"in": "header",
				"scopes": {
					"admin": "Accès administrateur pour gérer les films",
					"user": "Accès utilisateur standard"
				},
				"authorizationUrl": "http://swagger.io/api/oauth/dialog"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/user": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMIndexResponse_IUser_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page d'utilisateurs.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMCreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouvel utilisateur",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreate"
							}
						}
					}
				}
			}
		},
		"/user/{userId}": {
			"get": {
				"operationId": "ReadUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"description": "Récupérer une utilisateur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un utilisateur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMDeleteResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un utilisateur",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/film": {
			"get": {
				"operationId": "GetFilms",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMIndexResponse_IFilm_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page de film.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "CreateFilm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMCreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouveau film",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IFilmCreate"
							}
						}
					}
				}
			}
		},
		"/film/{filmId}": {
			"get": {
				"operationId": "ReadFilm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFilm"
								}
							}
						}
					}
				},
				"description": "Récupérer un film avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateFilm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un film avec le ID passé dans le URL",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IFilmUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFilm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMDeleteResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un film",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/film/{filmId}/actor": {
			"get": {
				"operationId": "GetActorsByFilm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IActor"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "🔹 Récupérer tous les acteurs d'un film.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/film/{filmId}/actor/{actorId}": {
			"get": {
				"operationId": "GetActorByFilm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IActor"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "🔹 Récupérer un acteur spécifique d'un film.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "actorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "AddActorToFilm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "🔹 Ajouter un acteur à un film (ADMIN uniquement).",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "actorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveActorFromFilm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "🔹 Supprimer un acteur d'un film (ADMIN uniquement).",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "filmId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "actorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/actor": {
			"get": {
				"operationId": "GetActors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMIndexResponse_IActor_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page d'acteur.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "CreateActor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMCreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouvel acteur",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IActorCreate"
							}
						}
					}
				}
			}
		},
		"/actor/{actorId}": {
			"get": {
				"operationId": "ReadActor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IActor"
								}
							}
						}
					}
				},
				"description": "Récupérer un acteur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "actorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateActor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un acteur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "actorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IActorUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteActor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORMDeleteResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un acteur",
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "actorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}